Clarification work
changed "events" to "pools" where they made sense
